patents/pyproject.toml
[project]
name = "patents"
version = "0.1.0"
description = "Patents data pipeline for SYWI platform"
requires-python = ">=3.11"
dependencies = [
    "dagster",
    "dagster-postgres",
    "pandas",
    "duckdb",
    "sywi-core",
    # BigQuery
    "google-cloud-bigquery",
    "db-dtypes",
    # LLM/Embeddings
    "dspy",
    "openai",
    # Graph databases
    "falkordb",
    "neo4j",
    # Data processing
    "numpy",
    # Visualization
    "matplotlib",
    # HTTP requests (for metadata fetching)
    "requests",
]

[tool.uv.sources]
sywi-core = { path = "../sywi-core", editable = true }

[project.optional-dependencies]
dev = ["dagster-webserver"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/patents"]

[tool.dg]
directory_type = "project"

[tool.dg.project]
root_module = "patents"
code_location_target_module = "patents.definitions"
